+++ src/Compilador/tree/Multiply.java
@@ -9,7 +9,7 @@ public class Multiply extends Tree {
         right = r;
     }
 
-    public Integer eval(Environment e) {
+    public Double eval(Environment e) {
         return left.eval(e) * right.eval(e);
     }
     public void print() {
