+++ src/Compilador/tree/Tree.java
@@ -1,7 +1,7 @@
 package Compilador.tree;
 
 public abstract class Tree {
-    public abstract Integer eval(Environment e);
+    public abstract Double eval(Environment e);
     public abstract void print();
 }
